Create Insturumentation test of pinning problem (#59)
