+++ app/build.gradle
@@ -19,6 +19,17 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    sourceSets {
+        String testCommon = "src/testCommon/java"
+        test {
+            java.srcDir testCommon
+        }
+        androidTest {
+            java.srcDir testCommon
+        }
+    }
+
 }
 
 dependencies {
