+++ app/src/androidTest/java/io/github/karino2/kotlitex/RenderTreeBuilderInstrumentedTest.kt
@@ -37,4 +37,25 @@ class RenderTreeBuilderInstrumentedTest {
         assertTrue(actual.isNotEmpty())
     }
 
+    @Test
+    fun buildExpression_sqrt() {
+        val opt = Options(Style.DISPLAY)
+        val input = parse("\\sqrt{3}")
+        val actual = RenderTreeBuilder.buildExpression(input, opt, true)
+
+        assertEquals(1, actual.size)
+        val target = actual[0]
+
+        assertSpan(target) { ac(0) { ac(0){ ac(0){ac(1)
+        {
+            // what canvas-latex obtain.
+            // styleTop("-2.916095em")
+
+            styleTop("-3.62519em")
+        }
+        }}}}
+    }
+
+
+
 }
